paypal enhancement for apps
improved the functionality on the paypalECButton render format. now uses
showLoading.
Also made a small change to showLoading to handle changing the text in a
loading modal.

There was also a bug in showLoading that cause 'undefined' to be
appended to the loading id if no id was set on the element. a guid is
now used (in combination w/ the tag type as some indicator of the
originator).